+++ src/Console/Command/DbResetCommand.php
@@ -42,7 +42,7 @@ class DbResetCommand extends Command
          */
         $this->addOption('type', [
             'description' => 'Use sql or php file',
-            'default' => Config::read('App.schemaFormat') ?? Config::read('Schema.format'),
+            'default' => Config::read('App.schemaFormat')
         ]);
 
         $this->checkForDeprecations();