+++ src/Console/Command/DbSetupCommand.php
@@ -36,7 +36,7 @@ class DbSetupCommand extends Command
         ]);
         $this->addOption('type', [
             'description' => 'Use sql or php file',
-            'default' => Config::read('App.schemaFormat') ?? Config::read('Schema.format'),
+            'default' => Config::read('App.schemaFormat')
         ]);
         $this->checkForDeprecations();
     }