+++ CHANGELOG.md
@@ -1,9 +1,22 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+## [3.0.0] - 2020-07-08
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+### Changed
+
+- Changed OriginPHP framework minimum 3.0
+- Changed PHP minimum 7.3
+- Changed PHPUnit minimum 9.2
+
+- Changed `Cache::clear` to remove `.env.php`
+
+### Removed
+
+- Removed deprecation notices
+
+## [2.0.0]
+
+Skipped this version
 
 ## [1.4.3] - 2020-07-03